bg(api): fix broken tests
- fixed the tests which were failing due to the change in folder structure
- re-arrange tests codes to be more maintainable and locate them in two modules

[Finishes #168864344]
