+++ server/models/articlesArray.js
@@ -29,7 +29,7 @@ const articlesArray = {
       title: 'My first Article',
       article: 'This is the very beginning of my writing journey, Although it seems to be hard I will keep fighting, Thank you for reading hope to see you next time.',
       createdOn: 'Tue Oct 01 2019 15:59:28 GMT+0200 (Central Africa Time)',
-      authorId: 'dd7f21c1-b92c-4703-a6d9-3ec03eef4da9',
+      authorId: '538bdd77-38af-4928-9bb0-d02461c7da34',
       comments: [],
     },
     {
