+++ server/tests/test.spec.js
@@ -1,25 +1,18 @@
 import chai, { expect } from 'chai';
 import chaiHttp from 'chai-http';
 import app from '../server';
-import signup from './signup.spec';
-import signin from './signin.spec';
-import createArticle from './createArticle.spec';
-import editArticle from './editArticle.spec';
-import deleteArticle from './deleteArticle.spec';
-import addComment from './addComment.spec';
-import feeds from './feeds.spec';
-import specificArticle from './specificArticle.spec';
+import auth from './auth.spec';
+import articles from './articles.spec';
 
 chai.use(chaiHttp);
 
-describe('POST /api/v1/auth/signup', signup);
-describe('POST /api/v1/auth/signin', signin);
-describe('POST /api/v1/articles', createArticle);
-describe('PATCH /articles/<articleId>', editArticle);
-describe('PATCH /articles/<articleId>', deleteArticle);
-describe('POST /api/v1/articles/<artilceId>/comments', addComment);
-describe('POST /api/v1/feeds', feeds);
-describe('POST /api/v1/articles', specificArticle);
+describe('POST /api/v1/auth/signup', auth.signup);
+describe('POST /api/v1/auth/signin', auth.signin);
+describe('POST /api/v1/articles', articles.create);
+describe('PATCH /articles/<articleId>', articles.edit);
+describe('DELETE /articles/<articleId>', articles.delete);
+describe('POST /api/v1/articles/<artilceId>/comments', articles.addComment);
+describe('POST /api/v1/feeds', articles.feeds);
 
 describe('Other routes', () => {
   it('GET / should return a welcome message', (done) => {
